agent_tool/pyproject.toml
[project]
name = "agent-tool"  # 注意：这里是 kebab-case
version = "0.1.0"
description = "Agent tool using Google ADK"
requires-python = ">=3.12"
dependencies = [
    "google-adk[litellm]>=1.23.0",
    "litellm>=1.81.5",
    "utils",  # 引用工作区内的 utils 包
]

[tool.hatch.build.targets.wheel]
packages = ["src/agent_tool"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"